Fix model relationships, implement vote verify
Model relationships were cleaned up so that it's a nice chain of
one-to-manys: A poll has many choices, and a choice has many votes.

Added a dummy API for development purposes. Running it drops all data
from the DB and puts in one poll with two choices, and four fake votes.
This is useful for quickly testing stuff (like GET choice, vote
verification) without having to manually create polls and choices
through Postman (or similar). Note that **running this will drop all
existing data in the database!**

Vote verification was "sort of" implemented - the check is done
correctly, but the way we retrieve the IP address of the user needs to
stem from the request data itself, not to be provided by the user.
Otherwise, the IP address would be _way_ too easy to fake.

Also added some comments about things to fix/talk about in the future.